BranchStaff id from Integer to Long

+++ BackEnd/park_management/src/main/java/park/management/com/vn/entity/BranchStaff.java
@@ -23,7 +23,7 @@ public class BranchStaff extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private ParkBranch parkBranch;
 
-    public BranchStaff(Integer id) {
+    public BranchStaff(Long id) {
         this.id = id;
     }
 }
